+++ src/Helpers/sass2scss.js
@@ -3,12 +3,12 @@
  * For LABOR.digital
  */
 module.exports = function sass2scss(content) {
-	if(typeof content !== 'string') return content;
+	if (typeof content !== "string") return content;
 
 	let lines = content.split(/\r?\n/g);
 
 	// Add finisher line
-	lines.push('#');
+	lines.push("#");
 
 	// The last indend length to check for new blocks
 	let lastIndent = 0;
@@ -30,7 +30,7 @@ module.exports = function sass2scss(content) {
 
 	// Loop over lines
 	for (let rawLine of lines) {
-		const indent = rawLine.replace(/^(\s*)(.*?)$/g, '$1').length;
+		const indent = rawLine.replace(/^(\s*)(.*?)$/g, "$1").length;
 
 		// Check if we are inside a block comment
 		if (inBlockComment) {
@@ -40,21 +40,21 @@ module.exports = function sass2scss(content) {
 				lastIndent = lastIndentBeforeBlockComment;
 			} else {
 				// Loop comments around the logic
-				parsedLines.push({'text': '//' + rawLine, 'comment': '', 'indent': indent, 'empty': true});
+				parsedLines.push({"text": "//" + rawLine, "comment": "", "indent": indent, "empty": true});
 				continue;
 			}
 		}
 
 		// Try to parse out inline comments
-		let commentIndex = rawLine.indexOf(' //');
-		if(commentIndex === -1 && rawLine.charAt(1) === '/') commentIndex = 0;
+		let commentIndex = rawLine.indexOf(" //");
+		if (commentIndex === -1 && rawLine.charAt(1) === "/") commentIndex = 0;
 		let line = rawLine;
-		let comment = '';
+		let comment = "";
 		if (commentIndex !== -1) {
 			comment = line.slice(commentIndex);
 			line = line.slice(0, commentIndex);
 		}
-		line = {'text': line, 'comment': comment, 'indent': indent};
+		line = {"text": line, "comment": comment, "indent": indent};
 
 		// Check if line is empty
 		if (line.text.trim().length === 0) {
@@ -67,7 +67,7 @@ module.exports = function sass2scss(content) {
 		if (line.text.match(/(?:^|\s)\/\*/)) {
 			inBlockComment = true;
 			line.empty = true;
-			line.text = '//' + line.text;
+			line.text = "//" + line.text;
 			parsedLines.push(line);
 			lastIndentBeforeBlockComment = lastIndent;
 			lastIndent = indent;
@@ -76,16 +76,16 @@ module.exports = function sass2scss(content) {
 
 		// Check if we are deeper than before
 		if (indent > lastIndent) {
-			if(indentHistory.indexOf(indent) === -1)
+			if (indentHistory.indexOf(indent) === -1)
 				indentHistory.push(indent);
 			// Update last line
-			parsedLines[lastLineWithContent].text += '{';
+			parsedLines[lastLineWithContent].text += "{";
 		}
 
 		// Check if we are higher than before
 		else if (lastIndent > indent) {
 			let layers = 0;
-			for(let i = indentHistory.length - 1; i >= 0 ; i--){
+			for (let i = indentHistory.length - 1; i >= 0; i--) {
 				const knownIndent = indentHistory[i];
 				if (knownIndent > indent) {
 					layers++;
@@ -93,14 +93,25 @@ module.exports = function sass2scss(content) {
 				}
 				break;
 			}
+
+			// Try to fix empty selectors on the previous line
+			let skip = false;
+			if (parsedLines[lastLineWithContent].text.match(/^\s*[^@+:,{()\n]*?$/g)) {
+				if (parsedLines[lastLineWithContent].text.trim() !== "" &&
+					!parsedLines[lastLineWithContent].text.match(/^\/\//)) {
+					parsedLines[lastLineWithContent].text += "{}";
+					skip = true;
+				}
+			}
+
 			indentHistory = indentHistory.slice(0, indentHistory.length - layers);
-			parsedLines[lastLineWithContent].text += '}'.repeat(layers);
+			parsedLines[lastLineWithContent].text += "}".repeat(layers);
 		}
 
 		lastIndent = indent;
 
 		parsedLines.push(line);
-		lastLineWithContent = parsedLines.length -1;
+		lastLineWithContent = parsedLines.length - 1;
 	}
 
 	// Drop finisher line
@@ -109,32 +120,30 @@ module.exports = function sass2scss(content) {
 	// Rebuild parsed lines into a list of scss lines
 	lines = [];
 	for (const [index, line] of parsedLines.entries()) {
-		if(line.empty !== true){
+		if (line.empty !== true) {
 			// Try to find empty selectors which may break the process
-			if(line.text.indexOf(':') === -1 && line.text.match(/^\s*[^:+=@\n]*?[^\n{,]$/)){
-				if(!parsedLines[index + 1] || parsedLines[index + 1].indent === line.indent){
-					line.text += '{}';
-				}
+			if (line.text.indexOf(":") === -1 && line.text.match(/^\s*[^:+=@\n{}]*?[^\n{,]$/)) {
+				line.text += "{}";
 			}
 
 			// Add semicolon to line if required
 			if (line.text.match(/[^,{](?:\s)?$/)) {
-				line.text += ';';
+				line.text += ";";
 			}
 		}
 
-		lines.push(line.text + ' ' + line.comment);
+		lines.push(line.text + " " + line.comment);
 	}
 
-	let output = lines.join('\r\n');
+	let output = lines.join("\r\n");
 
 	// Replace mixin definitions
 	output = output.replace(/(^\s*)([+=])(\s*)/gm, (a, before, char, after) => {
-		return before + (char === '+' ? '@include ' : '@mixin ') + after;
+		return before + (char === "+" ? "@include " : "@mixin ") + after;
 	});
 
 	// Remove all semicolons in front of an "else"
-	output = output.replace(/(});+([^;]*?@else)/gm, '$1$2');
+	output = output.replace(/(});+([^;]*?@else)/gm, "$1$2");
 
 	// Done
 	return output;